[SMA-230] Resolve cyclic dependencies
