+++ evoleq/src/commonMain/kotlin/org/evoleq/permission/combine-contexts.kt
@@ -1,0 +1,3 @@
+package org.evoleq.permission
+
+fun combine(vararg contexts: String): String = contexts.joinToString("/") { it }
