+++ evoleq/src/commonMain/kotlin/org/evoleq/permission/empty-context.kt
@@ -1,0 +1,12 @@
+package org.evoleq.permission
+
+import org.evoleq.value.StringValueWithDescription
+
+object EmptyContext:  StringValueWithDescription {
+    override val value = Value.EMPTY
+    override val description = ""
+}
+
+object Value {
+    const val EMPTY = "EMPTY"
+}
