+++ solawi-bid-api-data/src/commonMain/kotlin/org/solyton/solawi/bid/module/application/README.md
@@ -1,0 +1,2 @@
+# Application Management Module
+
