+++ solawi-bid-api-data/src/commonMain/kotlin/org/solyton/solawi/bid/module/application/permission/values.kt
@@ -1,0 +1,6 @@
+package org.solyton.solawi.bid.module.application.permission
+
+object Value {
+    const val ORGANIZATION = "ORGANIZATION"
+    const val APPLICATION = "APPLICATION"
+}
