+++ solawi-bid-api-data/src/commonMain/kotlin/org/solyton/solawi/bid/module/bid/permission/values.kt
@@ -1,0 +1,8 @@
+package org.solyton.solawi.bid.module.bid.permission
+
+
+object Value {
+    const val MANAGEMENT = "MANAGEMENT"
+    const val AUCTION = "AUCTION"
+    const val ORGANIZATION = "ORGANIZATION"
+}
