+++ solawi-bid-api-data/src/commonMain/kotlin/org/solyton/solawi/bid/module/user/permission/values.kt
@@ -1,0 +1,7 @@
+package org.solyton.solawi.bid.module.user.permission
+
+
+object Value {
+    const val ORGANIZATION = "ORGANIZATION"
+    const val MANAGEMENT = "MANAGEMENT"
+}
